@@ -1,9 +1,9 @@
 package com.cleanup.todoc.modelpersistance;
 
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.PrimaryKey;
-import android.support.annotation.ColorInt;
-import android.support.annotation.NonNull;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
+import androidx.annotation.ColorInt;
+import androidx.annotation.NonNull;
 
 /**
  * <p>Models for project in which tasks are included.</p>
