@@ -1,8 +1,8 @@
 package com.cleanup.todoc.ui.injections;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.annotation.NonNull;
 
 import com.cleanup.todoc.read.businesslogic.usecases.RetrieveTasksByProject;
 import com.cleanup.todoc.ui.TaskViewModel;
